requirements.txt
streamlit
pandas
xlsxwriter
matplotlib
Pillow